requirements.txt
numpy==2.3.2
pyfluidsynth==1.3.4
pygame==2.6.1
python-rtmidi==1.5.8
Flask==2.3.3
Flask-SocketIO==5.3.6
python-socketio==5.9.0
eventlet==0.33.3
psutil==5.9.0
